pyproject.toml
[project]
name = "SkillBridge"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "SeungminAhn", email = "smahn9123@gmail.com"},
]
dependencies = [
    "asgiref==3.8.1",
    "django==5.1.6",
    "pillow==11.1.0",
    "sqlparse==0.5.3",
    "tzdata==2025.1",
    "social-auth-app-django>=5.4.3",
    "django-allauth>=65.4.1",
    "django-widget-tweaks>=1.5.0",
    "python-dotenv>=1.0.1",
    "reportlab>=4.3.1",
    "iamport-rest-client>=0.9.0",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "flake8>=7.1.2",
    "pre-commit>=4.1.0",
    "black>=25.1.0",
]
